% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEDashboard.R
\name{DEDashboard}
\alias{DEDashboard}
\title{DEDashboard}
\usage{
DEDashboard(object, assay, alpha = 0.05, l2fc = 1, num = NULL, den = NULL)
}
\arguments{
\item{object}{A \code{geneda} object containing DEGs in the DEGs slot.}

\item{assay}{Character. Name of the assay to visualize.}

\item{alpha}{Numeric. Significance threshold for highlighting p-values and DEGs (default 0.05).}

\item{l2fc}{Numeric. Log2 fold-change threshold for MA and volcano plots.}

\item{num}{Character. Name of the numerator group for volcano plot labeling.}

\item{den}{Character. Name of the denominator group for volcano plot labeling.}
}
\value{
A \code{patchwork} object combining plots.
}
\description{
Generate a multi-panel differential expression dashboard for a single assay in a \code{geneda} object.
The dashboard includes:
\enumerate{
\item P-value histogram
\item MA plot
\item Volcano plot
}
}
