% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{DEGs}
\alias{DEGs}
\title{Access DEGs container}
\usage{
DEGs(object)
}
\arguments{
\item{object}{A \code{geneda} object}
}
\value{
List with \code{unfiltered} and \code{filtered} data.frames
}
\description{
Access DEGs container
}
