% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram}
\alias{dendrogram}
\title{dendrogram}
\usage{
dendrogram(MAT, META, FEATURE, COLORS, OUTPUT)
}
\arguments{
\item{MAT}{A data matrix where rows are features and columns are samples}

\item{META}{Metadata as a dataframe consisting of the samples provided in MAT. Samples should be rows and columns are metadata features}

\item{FEATURE}{A specific metadata feature of which the dendrogram branches will be colored by}

\item{COLORS}{A named color vector with one entry for every unique level in your FEATURE}

\item{OUTPUT}{A directory path to where the output should be saved. Path needs to end in a "/"}
}
\value{
A sample dendrogram in tiff format at the specified output path
}
\description{
Plot sample dendrogram colored by metadata feature.
}
\examples{
color_vector <- c("A" = "red", "B" = "blue")
output_folder <- c("Path/to/outputs/")
dendrogram(myDataMatrix, metadataDF, "some_column", color_vector, output_folder)
}
