% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{DimReduction}
\alias{DimReduction}
\title{Access dimensional reduction results list}
\usage{
DimReduction(object)
}
\arguments{
\item{object}{A \code{geneda} object}
}
\value{
List with \code{Loadings}, \code{Eigenvectors}, \code{percent_var}
}
\description{
Access dimensional reduction results list
}
