% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigencorr_MGX.R
\name{eigencorr_MGX}
\alias{eigencorr_MGX}
\title{eigencorr_MGX}
\usage{
eigencorr_MGX(mgx, metadata, num_mds = 10, OUTPUT)
}
\arguments{
\item{OUTPUT}{directory path to where the output should be saved. Path needs to end in a "/"}

\item{MAT}{A data matrix where rows are features and columns are samples}

\item{META}{Metadata as a dataframe consisting of the samples provided in MAT. Samples should be rows and columns are metadata features. If you only want to explore certain metadata features, create a subset of your metadata to only include columns of interest.}

\item{NUM_mds}{Number of NMDS to use.}
}
\value{
A list consisting of 3 slots: cor_matrix which holds correlation results, pval_matrix which holds p-value results, and stars which holds significance stars for plotting
}
\description{
Calculate Eigen-correlations for MGX data to correlate axes of variation with Bray-Curtis distances. For continuous variables, the function uses Pearson correlation. Pvalues follow this convention: p < 0.001, p < 0.01, p < 0.05 = three stars, two stars, one star, respectively
}
