% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEigen.R
\name{extractEigen}
\alias{extractEigen}
\title{extractEigen}
\usage{
extractEigen(object, component)
}
\arguments{
\item{object}{A \code{geneda} object containing PCA results in the \code{DimReduction} slot}

\item{component}{Principal component to use (i.e., PC1)}
}
\value{
A data-frame containing gene loading information and percent variation per gene
}
\description{
Extract gene-level Eigen-vectors for PCs of interest and calculate percent variance.
}
\examples{
\donttest{
extractEigen(object = obj, component = "PC1")
}
}
