% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLoadings.R
\name{extractLoadings}
\alias{extractLoadings}
\title{extractLoadings}
\usage{
extractLoadings(variance, pcRes, component, nfeatures)
}
\arguments{
\item{variance}{A vector of gene variances (can calculate using RGenEDA::plotVariance)}

\item{pcRes}{Output list from RGenEDA::generatePCs}

\item{component}{Principal component to use (i.e., PC1)}

\item{nfeatures}{Number of features principal components were generated on}
}
\value{
A data-frame containing gene loading information and percent variation per gene
}
\description{
Extract gene loadings for PCs of interest and calculate percent variance at the per-gene level
}
