% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Functions.R
\name{FindHVDEGs}
\alias{FindHVDEGs}
\title{Find DEGs Overlapping With HVGs}
\usage{
FindHVDEGs(object, assay, direction = c("positive", "negative", "both"))
}
\arguments{
\item{object}{A \code{geneda} object with \code{DEGs$DEG} set}

\item{assay}{The DEG slot to pull results from (typically a result of \code{FilterDEGs})}

\item{direction}{One of positive, negative, or both (indicating the direction
of DEG fold change to intersect with HVGs)}
}
\description{
Looks for DEGs after filtering that overlap with HVGs.
}
