% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{FindVariableFeatures}
\alias{FindVariableFeatures}
\title{Find and store highly variable genes (HVGs)}
\usage{
FindVariableFeatures(object, nfeatures)
}
\arguments{
\item{object}{A \code{geneda} object}

\item{nfeatures}{Number of HVGs to store}
}
\value{
Updated \code{geneda} object
}
\description{
Computes per-feature variance on \code{normalized(object)}, ranks genes by
decreasing variance, and stores the top \code{nfeatures} gene IDs into the \code{HVGs}
slot. Returns the updated object.
}
