% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Functions.R
\name{generatePCs}
\alias{generatePCs}
\title{generatePCs}
\usage{
generatePCs(mat, vars, nFeatures)
}
\arguments{
\item{mat}{A data matrix where rows are features and columns are samples}

\item{vars}{A vector of gene variances (can calculate using RGenEDA::plotVariance)}

\item{nFeatures}{Number of top features to generate principal components on.}
}
\value{
A list consisting of 3 slots: Loadings, Eigenvectors, and percent_var
}
\description{
Generate principal component analysis data that can be used in downstream analyses.
}
