% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{getCounts}
\alias{getCounts}
\title{Access counts matrix (avoid name clash with DESeq2::counts)}
\usage{
getCounts(object)
}
\arguments{
\item{object}{A \code{geneda} object}
}
\value{
Matrix or NULL
}
\description{
Access counts matrix (avoid name clash with DESeq2::counts)
}
