% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getMetadata}
\alias{getMetadata}
\title{Access metadata (avoid name clash with S4 generics)}
\usage{
getMetadata(object)
}
\arguments{
\item{object}{A \code{geneda} object}
}
\value{
data.frame
}
\description{
Access metadata (avoid name clash with S4 generics)
}
