% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getNormalized}
\alias{getNormalized}
\title{Access normalized matrix (avoid name clash with generics)}
\usage{
getNormalized(object)
}
\arguments{
\item{object}{A \code{geneda} object}
}
\value{
Matrix
}
\description{
Access normalized matrix (avoid name clash with generics)
}
