% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAplot.R
\name{MAplot}
\alias{MAplot}
\title{MAplot}
\usage{
MAplot(
  results,
  numerator,
  refLevel,
  log2FC_thresh,
  padj_thresh,
  title,
  figDir,
  figName
)
}
\arguments{
\item{results}{A dataframe of bulk RNA-Seq differential expression results from DESeq2}

\item{numerator}{Numerator level name}

\item{refLevel}{Reference level name}

\item{log2FC_thresh}{Absolute value of log2FC magnitude threshold}

\item{padj_thresh}{Adjusted p-value significance threshold}

\item{title}{Plot title}

\item{figDir}{Output directory ending in "/"}

\item{figName}{Output figure name with .png extension}
}
\value{
An output figure and ggplot object
}
\description{
Generate publicatrion quality MA plot for bulk RNA-Seq differential expression results
}
