% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{normalized}
\alias{normalized}
\title{Access normalized matrix}
\usage{
normalized(object)
}
\arguments{
\item{object}{A \code{geneda} object}
}
\value{
Matrix
}
\description{
Access normalized matrix
}
