% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMA.R
\name{PlotMA}
\alias{PlotMA}
\title{PlotMA}
\usage{
PlotMA(object, alpha, fc, title = NULL)
}
\arguments{
\item{object}{A \code{geneda} object containing \code{DEGs} from \code{SetDEGs} method}

\item{alpha}{Threshold for adjusted p-values (padj column from \code{DESeq2})}

\item{fc}{Absolute value log2Fold-change magnitude threshold (log2FoldChange column)}

\item{title}{Optional character vector of what plot should be titled.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot MA plot from a \code{geneda} object. Requires that you have ran
\code{SetDEGs} on your object following differential expression in \code{DESeq2}.
}
\examples{
\donttest{
ec <- eigencorr(obj, num_pcs = 5)
ec$plot
}
}
