% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{PlotPCA}
\alias{PlotPCA}
\title{Plot PCA results from a GenEDA object}
\usage{
PlotPCA(
  object,
  x = 1,
  y = 2,
  color_by,
  colors = NULL,
  split_by = NULL,
  shape_by = NULL,
  return_data = FALSE
)
}
\arguments{
\item{object}{A \code{GenEDA} object containing PCA results in the \code{DimReduction} slot.}

\item{x}{Numeric or character. The principal component to plot on the x-axis (e.g., 1 or "PC1").}

\item{y}{Numeric or character. The principal component to plot on the y-axis (e.g., 2 or "PC2").}

\item{color_by}{Character. Column name in the metadata used to color points.}

\item{colors}{Vector. Custom colors to use for plotting}

\item{split_by}{Character (optional). Column name in metadata used for faceting (creates separate panels).}

\item{shape_by}{Character (optional). Column name in metadata used to control point shape.}

\item{return_data}{Logical (default = FALSE), whether or not to return pca dataframe for more custom plotting.}
}
\value{
A \code{ggplot} object displaying the PCA scatter plot, or a list of pca_df and plot if \code{return_data = TRUE}
}
\description{
Visualize principal component analysis (PCA) results stored within a
\code{GenEDA} object. This function extracts PCA data via \code{\link{ExtractPCA}}
and provides a flexible ggplot2-based visualization interface.
}
\examples{
\dontrun{
p <- PlotPCA(obj, x = 1, y = 2, color_by = "condition",
     colors = c("untreated" = "red", "treated" = "blue"),
     split_by = "library")
p
}
}
