% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotScree.R
\name{PlotScree}
\alias{PlotScree}
\title{PlotScree}
\usage{
PlotScree(object)
}
\arguments{
\item{object}{A \code{geneda} object containing PCA results in the DimReduction slot.}
}
\value{
A ggplot2 object
}
\description{
Generate a scree plot
}
