% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariance.R
\name{plotVariance}
\alias{plotVariance}
\title{plotVariance}
\usage{
plotVariance(MAT, OUTPUT, LOG = FALSE, nFeaturesDrop = NULL)
}
\arguments{
\item{MAT}{A data matrix where rows are features and columns are samples}

\item{OUTPUT}{A directory path to where the output should be saved. Path needs to end in a "/"}

\item{LOG}{Log transform variance. Default = FALSE}

\item{nFeaturesDrop}{Number of top most variable features to drop. Useful if top genes are extremely variable. Default = NULL}
}
\value{
A vector of variances named by gene. A gene by variance plot in tiff format at the specified output location.
}
\description{
Analyze gene variance to determine thresholds for PCA. Plot gene by variance plot.
}
\examples{
output_folder <- c("Path/to/outputs/")
# Plot variance as it
plotVariance(matrix, output_folder)
# Plot logged transformed variance
plotVariance(matrix, output_folder, LOG = TRUE)
# Plot variance with top 3 features dropped
plotVariance(matrix, output_folder, nFeaturesDrop = 3)
# If running PCA, save outputs to variable to access the features
vars <- plotVariance(matrix, output_folder)
}
