% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVolcano.R
\name{PlotVolcano}
\alias{PlotVolcano}
\title{PlotVolcano}
\usage{
PlotVolcano(object, alpha, fc, den, num, title = NULL)
}
\arguments{
\item{object}{A \code{geneda} object containing \code{DEGs} from \code{SetDEGs} method}

\item{alpha}{Threshold for adjusted p-values (padj column from \code{DESeq2})}

\item{fc}{Absolute value log2Fold-change magnitude threshold (log2FoldChange column)}

\item{den}{Denominator (reference level for comparison, num vs. den)}

\item{num}{Numerator (numerator level for comparison, num vs. den)}

\item{title}{Optional character vector of what plot should be titled.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot volcano plot from a \code{geneda} object. Requires that you have ran
\code{SetDEGs} on your object following differential expression in \code{DESeq2}.
}
\examples{
\donttest{
mock_norm <- matrix(rnorm(5000 * 6, mean = 0, sd = 2), nrow = 5000, ncol = 6)
colnames(mock_norm) <- paste0("Sample", 1:6)
rownames(mock_norm) <- paste0("Gene", 1:5000)

# Sample metadata
mock_meta <- data.frame(condition = c("A","B","A","B","A","B"),
row.names = colnames(mock_norm))

# Construct GenEDA object
obj <- GenEDA(normalized = mock_norm, metadata = mock_meta)

# make dummy DEG table
n <- 1000
genes <- paste0("Gene", 1:n)
log2FoldChange <- runif(n, min = -5, max = 5)
padj <- runif(n, min = 0, max = 1)
df <- data.frame(
 log2FoldChange = log2FoldChange,
 padj = padj,
 stringsAsFactors = FALSE)
rownames(df) <- genes
obj <- SetDEGs(obj, df)

PlotVolcano(obj, 0.05, 1, "Denominator Group", "Numerator Group", "Test")
}
}
