% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{SetDEGs}
\alias{SetDEGs}
\title{Set unfiltered DEGs on the object}
\usage{
SetDEGs(object, deg_table)
}
\arguments{
\item{object}{A \code{geneda} object}

\item{deg_table}{A data.frame of DESeq2-like results containing at least
columns \code{log2FoldChange} and \code{padj}}
}
\value{
Updated \code{geneda} object with \code{DEGs$unfiltered} set and \code{DEGs$filtered} cleared
}
\description{
Set unfiltered DEGs on the object
}
