% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,geneda-method}
\alias{show,geneda-method}
\title{Show method for geneda}
\usage{
\S4method{show}{geneda}(object)
}
\arguments{
\item{object}{A \code{geneda} object}
}
\description{
Show method for geneda
}
