% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnailKO.R
\docType{data}
\name{Snail1KO}
\alias{Snail1KO}
\title{Snail1KO dataset}
\format{
A list containing raw counts and metadata
\describe{
\item{rawCounts} {A dataframe of raw counts expression matrix}
\item{metadata} {A dataframe of associated metadata}
}
}
\source{
Maturi et al., 2018; ENA: ERP019920; E-MTAB-5244
}
\usage{
Snail1KO
}
\description{
Dataset from Maturi, et al. 2018, (ENA:ERP019920, E-MTAB-5244)
}
\keyword{datasets}
