% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{DEGs}
\alias{DEGs}
\title{Access DEGs container}
\usage{
DEGs(object, assay)
}
\arguments{
\item{object}{A \code{geneda} object}

\item{assay}{The name of the DEGs slot to return}
}
\value{
List with \code{DEG} (unfiltered data.frame) and optionally named filtered result sets
}
\description{
Access DEGs container
}
