% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceHeatmap.R
\name{distanceHeatmap}
\alias{distanceHeatmap}
\title{distanceHeatmap}
\usage{
distanceHeatmap(MAT, META, FEATURES = c(...), PALETTES, OUTPUT)
}
\arguments{
\item{MAT}{A data matrix where rows are features and columns are samples}

\item{META}{Metadata as a dataframe consisting of the samples provided in MAT. Samples should be rows and columns are metadata features}

\item{PALETTES}{A list where each list slot is a named vector of metadata levels and colors. Each feature present in FEATURE needs to have its own list slot}

\item{OUTPUT}{A directory path to where the output should be saved. Path needs to end in a "/"}

\item{FEATURE}{A vector of features to be included in the heatmap annotations}
}
\value{
A distance heatmap in png format at the specified output path
}
\description{
Calculate sample to sample distances using Euclidean distances. Create annotated heatmap to assess data trends.
}
\examples{
timepoint_colors <- c("Early" = "red", "Late" = "blue")
treatment_colors <- c("Mock" = "pink", "HSV" = "green")
color_list <- list(timepoint_colors, treatment_colors)
output_folder <- c("Path/to/outputs/")
dendrogram(myDataMatrix, metadataDF, c("Timepoint", "Treatment"), color_list, output_folder)
}
