% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigencorr.R
\name{eigencorr}
\alias{eigencorr}
\title{eigencorr}
\usage{
eigencorr(MAT, META, NUM_PCS = 10, OUTPUT)
}
\arguments{
\item{MAT}{A data matrix where rows are features and columns are samples}

\item{META}{Metadata as a dataframe consisting of the samples provided in MAT. Samples should be rows and columns are metadata features. If you only want to explore certain metadata features, create a subset of your metadata to only include columns of interest.}

\item{NUM_PCS}{Number of principal components to correlate.}

\item{OUTPUT}{directory path to where the output should be saved. Path needs to end in a "/"}
}
\value{
A list consisting of 3 slots: cor_matrix which holds correlation results, pval_matrix which holds p-value results, and stars which holds significance stars for plotting
}
\description{
Calculate Eigen-correlations to correlate axes of variation with principal components.For continuous variables, the function uses Pearson correlation. Pvalues follow this convention: p < 0.001, p < 0.01, p < 0.05 = three stars, two stars, one star, respectively
}
\examples{
# Set output directory
outputDir <- c("/Path/to/output/")
#Subset metadata to only factor of interest
metaSub <- metadata[,colnames(metadata) \%in\% c("Timepoint", "Treatment")]
ecs <- eigencorr(matrix, metadata, 3, outputDir)
}
