% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEigen.R
\name{extractEigen}
\alias{extractEigen}
\title{extractEigen}
\usage{
extractEigen(object, component)
}
\arguments{
\item{object}{A \code{geneda} object containing PCA results in the \code{DimReduction} slot}

\item{component}{Principal component to use (i.e., PC1)}
}
\value{
A data-frame containing gene loading information and percent variation per gene
}
\description{
Extract gene-level Eigen-vectors for PCs of interest and calculate percent variance.
}
\examples{
\donttest{
mock_norm <- matrix(rnorm(12, mean = 0, sd = 2), nrow = 4, ncol = 3)
colnames(mock_norm) <- paste0("Sample", 1:3)
rownames(mock_norm) <- paste0("Gene", 1:4)
mock_meta <- data.frame(condition = c("A","B","A"), row.names = colnames(mock_norm))
obj <- GenEDA(normalized = mock_norm, metadata = mock_meta)
extractEigen(object = obj, component = "PC1")
}
}
