% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{FilterDEGs}
\alias{FilterDEGs}
\title{Filter DEGs by padj and absolute log2FoldChange}
\usage{
FilterDEGs(object, padj_thresh = 0.05, log2FC_thresh = 1)
}
\arguments{
\item{object}{A \code{geneda} object with \code{DEGs$unfiltered} set}

\item{padj_thresh}{Adjusted p-value threshold (<=)}

\item{log2FC_thresh}{Absolute log2 fold change threshold (>=)}
}
\value{
Updated \code{geneda} object with \code{DEGs$filtered} set
}
\description{
Filter DEGs by padj and absolute log2FoldChange
}
