% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePCs.R
\name{generatePCs}
\alias{generatePCs}
\title{generatePCs}
\usage{
generatePCs(MAT, VARS, NFEATURES)
}
\arguments{
\item{MAT}{A data matrix where rows are features and columns are samples}

\item{VARS}{A vector of gene variances (can calculate using DACExploreR::plotVariance)}

\item{NFEATURES}{Number of top features to generate principal components on}
}
\value{
A list consisting of 3 slots: Loadings, Eigenvectors, and percent_var
}
\description{
Generate principal component analysis data that can be used in downstream analyses.
}
\examples{
# Calculate variance
vars <- DACExplorer::plotVariance(matrix, output_folder)
pcs <- plotVariance(matrix, vars, 2000)
}
