% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCountDist.R
\name{PlotCountDist}
\alias{PlotCountDist}
\title{PlotCountDist}
\usage{
PlotCountDist(object, split_by = NULL)
}
\arguments{
\item{object}{A \code{geneda} object containing normalized expression values in \verb{@normalized}
and sample metadata in \verb{@metadata}.}

\item{split_by}{Character. Optional column name from \code{object@metadata} used
for faceting (default: NULL, no faceting).}
}
\value{
A ggplot2 object displaying the distribution of counts per sample.
}
\description{
Generates violin plots with embedded boxplots showing the distribution
of normalized or VST-transformed RNA-Seq counts for each sample in a
\code{geneda} object. Optionally, plots can be faceted by a metadata variable.
}
\examples{
\dontrun{
# Basic plot without faceting
PlotCountDist(obj)

# Facet by a metadata variable "condition"
PlotCountDist(obj, split_by = "condition")
}

\donttest{
mock_norm <- matrix(rnorm(12, mean = 0, sd = 2), nrow = 4, ncol = 3)
colnames(mock_norm) <- paste0("Sample", 1:3)
rownames(mock_norm) <- paste0("Gene", 1:4)
mock_meta <- data.frame(condition = c("A","B","A"), row.names = colnames(mock_norm))
obj <- GenEDA(normalized = mock_norm, metadata = mock_meta)
PlotCountDist(obj)
}
}
