% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneda-class.R
\name{plotHVGVariance}
\alias{plotHVGVariance}
\title{plotHVGVariance}
\usage{
plotHVGVariance(object, transform = NULL, dropTopN = 0)
}
\arguments{
\item{object}{A \code{geneda} object created by \code{GenEDA()}.}

\item{transform}{One of c("log") to log-transform variance, or NULL for none. Default = NULL}

\item{dropTopN}{Number of top-most variable genes to drop prior to plotting. Default = 0}
}
\value{
A ggplot2 object visualizing the variance of HVGs
}
\description{
Plot variance for all genes (features) on a \code{geneda} object using ggplot2.
}
\examples{
# obj <- GenEDA(vsd, meta)
# p <- plotHVGVariance(obj, transform = "log")
# print(p)
}
