% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{SetDEGs}
\alias{SetDEGs}
\title{Set unfiltered DEGs on the object}
\usage{
SetDEGs(object, deg_table, assay)
}
\arguments{
\item{object}{A \code{geneda} object}

\item{deg_table}{A data.frame of DESeq2-like results containing at least
columns \code{log2FoldChange} and \code{padj}}

\item{assay}{What to name the DEG slot}
}
\value{
Updated \code{geneda} object with \code{DEGs$DEG} set
}
\description{
Set unfiltered DEGs on the object
}
